refactor(dashboard-tsr): split topo-canvas glyphs into topo-glyphs (#1392)
Move the ClickHouse/Keeper glyph subcomponents, path helpers, and brand
constants into topo-glyphs.tsx; topo-canvas now only orchestrates layout +
edges. TopoCanvas export unchanged.

Part of epic #1392
